python_producer/pyproject.toml
[project]
name = "python-producer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python-ng>=2.2.3",
]
